extract Toolbar component
